Extruder axis movement fix
This moves the logic controlling the active tool head out of the path
planner, up into the actual G-code handler.  When a G0/1 command
includes an E movement, the E axis is replaced in the path with the axis
for the selected tool.  This keeps each extruder position separate,
allowing for correct movement.  Previously, the movements would all be
aggregated into the E position, making it impossible to swap between
tools without zeroing the position on each change.