refactor Path and PathPlanner so they aren't friends
It's currently difficult to reason about or modify the path planner code
because Path declares PathPlanner as a friend class and PathPlanner mucks
about with the internals of Path. This commit just moves code around until
the friend declaration is no longer necessary and there's a much clearer
separation between Path and PathPlanner.

Larger changes include:
- many Path setters are replaced by Path::initialize
- many early Path calculations are moved from PathPlanner to
  Path::initialize
- PathPlanner::calculatePath is moved to Path::calculate
- PathPlanner::calculateSafeSpeed is moved to Path::calculateSafeSpeed
- stepper parameters are refactored into their own StepperPathParameters
  struct rather than being fields in Path
- a few unused flags and fields are removed
- Path has getters and setters for fields that were previously accessed
  directly
    - these also allow the stepper parameters to be invalidated more
      reliably
- a few tweaks to pass std::vectors by const reference wherever possible