divide paths so each buffer we send to the PRU is at most 1/4th of the available PRU memory - fixes #122
Currently, we split buffers of steps going to the PRU just enough to make them fit in the total PRU memory,
and we always split them into evenly sized pieces. This creates a worst case when a path has a number of steps
that is just above a multiple of the number of steps that will fit in the PRU memory. The specific example
that made this clear was a path with just over 32,700 steps when the PRU buffers hold about that much. In this case,
we split the path into two separate buffers for the PRU, but each buffer occupies just over half the PRU memory
and we can't queue them both at once. This means that we queue one of them, the PRU runs through it, the PRU
triggers an interrupt, and the steppers pause for a moment while we queue the second buffer. This pause happens in
the middle of the path when it's already up to speed, so it can lead to missed steps or extra steps if the steppers
can't stop instantly or accelerate back to cruising speed once the second buffer reaches the PRU.

To avoid this worst case, simply split buffers sent to the PRU so they fill at most 1/4th of the PRU's memory.
This means that even in the worst case, the PRU will have a buffer of steps to run while we load the next one.

Thanks to Thomas Matern for helping diagnose this problem and validating the fix.