disable path planner asserts in normal builds
Ideally, we'd always run with asserts enabled and we'd have some
normal process for dealing with crashes (tell the user they have
a crash dump and/or automagically upload it somewhere with their
permission). What actually happens today is that redeem hits an
assert and crashes, then the watchdog reboots the BBB before
the systemd-coredump has finished dumping the process (and that's
assuming the user has installed systemd-coredump - it's not there
in the default image).

Hopefully we'll figure out a better way to handle this, but for now,
let's avoid saddling users with problems they can't solve and turn
off production asserts. Building the path planner without the rest
of redeem will still enable asserts, so users that want to help
diagnose issues can still do so in that way.