comment path assert that fires due to minor optimizer bug
As it turns out, the optimizer can over-accelerate on very short
moves (a few steps). This is bad, but it's been the case for a very
long time and no one has noticed or cared. For now, comment the path
planner assert that shows this bug so we don't break printers. In
the long run, we need more safety checks in the optimizer to make
sure the speeds it sets are actually reachable.