remove min_buffered_move_time
A recent commit to the path planner neutered the effects of
min_buffered_move_time with the reasoning that paths need to move
from the optimizer to the PRU at the rate the PRU consumes them,
and no amount of fiddling with the path or PRU queues will change
that speed. The system is generally self-correcting - if the optimizer
can't keep up with the rate at which paths are consumed, paths will
be passed to the PRU unoptimized, which will slow the rate at which
paths can be consumed. The alternative is to stall the PRU while
waiting for the optimizer, which is much worse because it can damage
a print.

This commit finishes the job by removing minBufferedMoveTime in the
path planner, min_buffered_move_time in the Python wrapper, and
the matching configuration option.

It's also worth noting that if this commit is reverted, the comment
next to the config option is actually incorrect.