if no fans are configured for M106/M107, don't do anything
Fixes #131
The current implementation of M106/M107 assumes that if the user
didn't specify a fan to modify and no fans are specified in the
config, fan 0 should be modified. This is a bad idea because it
makes Redeem possibly turn off the user's hotend fan if they haven't
configured their cooling fan (or if they don't have one but their
slicer is issuing commands as if they do). Instead, if the user
hasn't configured M106/M107 to affect any fans, they won't do anything.