Add a VectorN type to replace heap-allocated std::vector<double>
std::vector<double> is a lovely type, but it's heap allocated,
variable length, and we have to loop through every element to
perform any operation. This commit replaces it with an internal
VectorN type that's a fixed-length in-place allocated array
of NUM_AXES doubles with appropriate operator overloads to simplify
usage.