PRU Stepper Firmware Rewrite
The cycle delays in the firmware for the PRU that runs the steppers
are calculated for the old TI steppers and don't work for the Trinamics.
This commit introduces support for PRU firmwares written in C and adds
a replacement firmware that's written in C. Additionally, stepper driver
delays are now determined based on the Replicape revision and written
into the config for the firmware.

This should make our speeds more accurate. Additionally, it makes
probing a bit quicker because the new firmware is faster at cancelling
moves.