Time Dilation Path Planner Partial Rewrite
In the current acceleration ramp path planner, we calculate and
recalculate parameters for a move, but don't actually calculate
the steps for a path until it needs to be queued up on the realtime
coprocessor. This works, but it assumes the steps themselves are
relatively quick to calculate. It also means that acceleration
and change in position must both be taken into account in the same
calculation.

In this partial rewrite, we instead split path calculations into
two phases. In the first phase, we calculate the steps needed to
move from the start point to the end point of a path, timing them
as if we were moving at a constant speed. The second phase occurs
just as the path is headed to the realtime coprocessor, and it
involves running the precalculated steps through a "time dilation"
function that spreads them apart at the start and end of the move
to create the necessary acceleration and deceleration. It can also
change the cruising speed if necessary.

This has little benefit for cartesians, although because the second
phase is cheaper to calculate, it probably improves hot-path latency.
The chief benefit of this split is that the same "time dilation"
function can be applied to any set of steps for any motion platform.

This path planner also adds a new set of kinematics for delta printers
that can calculate all the steps needed to move in a straight line
in cartesian space from start to finish, with the assumption that
the move takes place at a constant speed. With the two-phase approach,
that assumption is acceptable because the "time dilation" function
will handle adding acceleration and deceleration later. This ultimately
enables completely segment-free accurate delta motion.