Remove Hez
It's not entirely clear what role Hez played in the old delta kinematics
because they sometimes calculated a position and sometimes returned it
from globals. In the new kinematics, though, it doesn't work. This isn't
really a necessary part of the kinematics and doesn't help usability, so
this change just removes it.