Fix delta XY-plane movement check
When a delta move is Z-only, we can bypass the complicated delta
kinematics and instead calculate it as a simple linear move. However,
the origianl check for this condition was a floating point equality
check. This is sensitive and fragile. Instead, this change replaces it
with a check that all the motors are moving the same number of steps
in the same direction.