Rewrite probing to be axis-independent
The original probe routine involved counting how many steps needed
to be traveled on the Z-axis, counting how many steps were skipped
after the move reached an endstop, and subtracting to figure out the
probe distance. That's fine, but it assumes single-axis movement.

With this change, the path planner takes over calculating probe
distances so they can be axis-independent. The path planner stores
all the steps it sends to the PRU, then counts how many steps were
executed to reconstruct the machine's state when the move was canceled.
This allows probing to be consistent on any machine or any direction.