remove min_speed and begin rename from jerk to max_speed_jump
The min_speed settings have just been confusing people for a while now,
and we've planned to remove min_speed because 0.5 * jerk serves the same
purpose and rename jerk to max_speed_jump because we, like other 3D
printer firmwares, don't implement "jerk" in the mathematical sense.

This change removes min_speed and does all the renaming in the code, but
leaves the max_jerk_* values in the config. These will have to stay
around until we can automatically fix users' configs.