make G90, G91, M82, and M83 Marlin-compatible
With this change, the printer has four states:
- Absolute movement (all axes)
- Relative movement (all axes)
- Absolute movement with relative extrusion
- Relative movement with a hidden state that extrusion will remain
relative even after switching back to absolute

This is to maintain compatibility with Marlin behavior that OctoPrint
relies on where G91 sets all axes to relative, including extruders. The
only state that becomes impossible with this change is relative movement
with absolute extrusion, which seems a bit silly.