PathPlanner: loosen asserts to catch double-steps
The path planner contains asserts to crash Redeem if path calculations
require us to step a motor twice in the same step-interval. This
generally catches path planner bugs, but it can also catch moves
that are too fast. As it turns out, these also have a nasty corner
case for two high speed moves in quick succession where the end times
for the moves fall in the same step-interval as the last step.

This will need to be accompanied by changes to cap move speeds to keep
them within the max steps/sec (currently 200,000).