insert sync events between movements and non-movements
One of the inherent problems in Redeem's architecture is that the
PRU takes care of its own timing for movements but non-movement
commands are processed on the main CPU. The most obvious result
of this was that fans turned on and off long before the gcode said
they should because the CPU is always running ahead of the PRU and
path planner's buffers.

This change adds logic to recognize that some commands complete
"asynchronously" via the path planner and PRU, and that the CPU
needs to wait for those commands to complete at some points. By
choosing these points carefully, we can continue building up a
buffer of upcoming paths while still synchronizing when needed.
This means the fans turn on at the right moment without the printer
pausing.