carry endstops across blocks of stepper commands when probing
One longstanding issue with probing is that if an endstop only fires a
short pulse, the PRU will only skip steps for the duration of that
pulse. Normally this isn't an issue because the PRU manages to skip
every step in the move during the pulse, but for long enough probe
moves, there may be some steps that arrive late enough to the PRU that
it starts moving again when the endstop pulse ends.

With this change, probe moves now set a flag that causes the PRU to
carry blocked axes forward through blocks of steps, which resolves the
issue.