rework probing to eliminate race condition
In the existing implementation, the python thread queueuing a path
was racing against the planner dispatch thread - both waited for the
PRU to finish the move but the python thread also needed to wait for
the planner thread to compute the probe distance.

Resolve this race by passing the probe result from the planning thread
to the queueing thread via a std::promise.