Comment path planner assert until it's fixed
This has been breaking people's prints. It was commented out under the old path planner - I uncommented it during the rewrite because I hope to address it. At this point I understand the underlying issue, but it's a challenge to fix because it's due to conflicting assumptions in how the optimizer is implemented.