Allow changes to TargetTemperature while waiting
This allows the user to alter temperature targets
without requiring the previous target to be reached

While the printer is waiting for heaters, M116 is the active gcode. Normally, this is a synchronous action.

However, M108 is allowed to be performed to allow the command to be aborted.

This change also allows M104 (Set extruder target temperature) and M140 (Set bed target temperature) to be executed while M116 remains active.