Revert M19 change of stepper class
The current testing of many functions is done
within the context of an instance of the
complete application (Redeem).

When this is generated, objects related to the
hardware are constructed on the basis of the
actual hardware found.

At the present, instances which were unable to
identify a proper Replicape configure as if
they are a “0B3A” version. The steppers in that
version are unchanged from the “00B3” version.

Thus the CI tests need to be based on that
assumed version until a testing proxy is
developed.