remove C++ dependencies that aren't needed for testing
When I first wrote the CMakeFile, I pulled in Python and SWIG. Python
was there because we needed its headers to drop and acquire the GIL from
C++, but I later removed that need by making SWIG handle the GIL. SWIG
was there because I thought we might eventually use CMake to build the
path planner from setup.py. This turned out to be more trouble than it's
worth.

Neither of these are actually needed when building the C++ unittests.