Apply Probe Z offset to G30 (#169)
* Apply Probe Z offset to G30

G30 previously took the X and Y probe offsets into account when
probing a given position. This allow getting bed information for
where the hotend should be instead of measuring at an offset
position.

This takes the Z offset into account when returning the probe
information. This also removes the offset from being indirectly
applied with the G33 commands.

* Fix formatting for changes to G30 command

Applies formatting rules to previous code changes

* Update test code for G30 and G33

The G30 and G33 commands have been updated in their handling of
the probe offset_z config parameter. This updates the test code
to test the changes.