Convert versioning to Released Branch Scheme
When a branch gets its FIRST release, this commit
should be the last commit on the branch and the
commit should be tagged “vN.N.0”

Future tags, “vN.N.1”, “vN.N.2”, etc. do not require
an additional commit.

However, a “base_N.M” tag will be needed at the start
of any new development that is not a part of the “N.N”
series.